+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Add a Group membership assignment quick form #723](https://github.com/farmOS/farmOS/pull/723)
 - [farmOS Setup Menu #706](https://github.com/farmOS/farmOS/pull/706)
 - [Issue #3354935: Configurable quick forms](https://www.drupal.org/project/farm/issues/3354935)
+- [Add an Account Admin role with permission to administer users and assign managed roles #714](https://github.com/farmOS/farmOS/pull/714)
 
 ### Changed